Update 2.pptx
In the future space crafts that will travel to and beyond the moon so it faces high risk so there should be some kind of safety to help the spacecraft  in the future missions. Our robot moves in an orbit fixed around the spacecraft to capture photos and detect any defects in the body of the spacecraft or any body approaching it, and when the sensor detects any defects it decreases its speed.